Moving vertices functions to its own module
